fix: export redis config getter to function
