+++ tests/integration/testContainerConfig.ts
@@ -3,6 +3,7 @@ import jsLogger from '@map-colonies/js-logger';
 import { RedisOptions } from 'ioredis';
 import { Registry } from 'prom-client';
 import { getConfig, initConfig } from '@src/common/config';
+import { getRedisConfig } from '@src/schema/utils/redisConfig';
 import { getSchemas } from '../../src/schema/providers/schemaLoader';
 import { REDIS_SYMBOL, SERVICES } from '../../src/common/constants';
 import { createConnection } from '../../src/common/db';
@@ -33,27 +34,7 @@ export const registerTestValues = async (params?: { appConfig?: IApplication; re
 
   const schemas = await getSchemas(container);
 
-  const redisConfig = config.get('db.redis');
-
-  if (!redisConfig) {
-    return undefined;
-  }
-
-  const { prefix, connectTimeoutMs, tls, ...rest } = redisConfig;
-  const usedPrefix = prefix ?? '';
-
-  let tlsOptions = undefined;
-  if (tls.enabled) {
-    const { enabled, ...tlsCerts } = tls;
-    tlsOptions = tlsCerts;
-  }
-
-  const redisConnection = await createConnection({
-    ...rest,
-    keyPrefix: usedPrefix,
-    connectTimeout: connectTimeoutMs,
-    ...(tlsOptions && { tls: tlsOptions }),
-  });
+  const redisConnection = await createConnection(getRedisConfig());
 
   container.register(SERVICES.SCHEMAS, { useValue: schemas });
   container.register(REDIS_SYMBOL, { useValue: redisConnection });
